+++ src/main/java/com/httm/WebBanThuocTay/model/HinhAnhThuoc.java
@@ -1,0 +1,41 @@
+package com.httm.WebBanThuocTay.model;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Embeddable;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Entity
+@Table(name = "HinhAnhThuoc")
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Embeddable
+public class HinhAnhThuoc {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "idHinhAnh")
+    private Integer id;
+
+    @Column(name = "hinhAnh", unique = true)
+    private String hinhAnh;
+    //Một hình ảnh cho một sản phẩm thuốc
+    @ManyToOne
+    @JoinColumn(name = "idThuoc")
+    @EqualsAndHashCode.Exclude
+    @ToString.Exclude
+    private Thuoc thuoc;
+
+}
