+++ src/main/resources/application.properties
@@ -6,6 +6,7 @@ spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.properties.hibernate.dialect= org.hibernate.dialect.SQLServer2008Dialect
 spring.jpa.hibernate.ddl-auto =none
+spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
 server.error.include-message= always
