+++ src/main/resources/templates/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Welcome to tiem thuoc tay</title>
+</head>
+<body>
+    <h1>Danh sach thuoc</h1>
+</body>
+</html>
